requirements.txt
Flask==2.0.2
flask-cors==3.0.10
flask-restx==0.5.1
pinecone-client==2.1.0
langchain-openai==0.0.8  # Specify the desired version
azure-identity==1.7.0
python-dotenv==0.19.2
colorama==0.4.4
pandas==1.3.3
faiss-cpu==1.7.1.post2
numpy==1.21.2
markdown==3.3.4